block: Move rpc methods to static constructor
