+++ packages/block/src/from-rpc.ts
@@ -45,6 +45,7 @@ function normalizeTxParams(_txParams: any) {
  * @param blockParams - Ethereum JSON RPC of block (eth_getBlockByNumber)
  * @param uncles - Optional list of Ethereum JSON RPC of uncles (eth_getUncleByBlockHashAndIndex)
  * @param options - An object describing the blockchain
+ * @deprecated
  */
 export function blockFromRpc(
   blockParams: JsonRpcBlock,
@@ -64,50 +65,4 @@ export function blockFromRpc(
   const uncleHeaders = uncles.map((uh) => blockHeaderFromRpc(uh, options))
 
   return Block.fromBlockData({ header, transactions, uncleHeaders }, options)
-}
-
-/**
- *  method to retrieve a block from the provider to use in the VM
- * @param provider an Ethers JsonRPCProvider
- * @param blockTag block hash or block number to be run
- * @param common Common instance used in VM
- * @returns the block specified by `blockTag`
- */
-export const getBlockFromProvider = async (
-  provider: ethers.providers.JsonRpcProvider,
-  blockTag: string | bigint,
-  common: Common
-) => {
-  let blockData
-  if (typeof blockTag === 'string' && blockTag.length === 66) {
-    blockData = await provider.send('eth_getBlockByHash', [blockTag, true])
-  } else if (typeof blockTag === 'bigint') {
-    blockData = await provider.send('eth_getBlockByNumber', [bigIntToHex(blockTag), true])
-  } else if (
-    isHexPrefixed(blockTag) ||
-    blockTag === 'latest' ||
-    blockTag === 'earliest' ||
-    blockTag === 'pending'
-  ) {
-    blockData = await provider.send('eth_getBlockByNumber', [blockTag, true])
-  } else {
-    throw new Error(
-      `expected blockTag to be block hash, bigint, hex prefixed string, or earliest/latest/pending; got ${blockTag}`
-    )
-  }
-
-  const uncleHeaders = []
-  if (blockData.uncles.length > 0) {
-    for (let x = 0; x < blockData.uncles.length; x++) {
-      const headerData = await provider.send('eth_getUncleByBlockHashAndIndex', [
-        blockData.hash,
-        intToHex(x),
-      ])
-      uncleHeaders.push(headerData)
-    }
-  }
-
-  return blockFromRpc(blockData, uncleHeaders, {
-    common,
-  })
 }
